More work on flow tests
