+++ lib/flow.js
@@ -6,7 +6,7 @@
  * @returns {Function} Composed function.
  */
 function flow(...funcs) {
-
+  return () => {};
 }
 
 module.exports = flow;
